requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.5
cryptography==41.0.2
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
humps==0.2.2
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
pycparser==2.21
pydantic==1.10.4
PyJWT==2.7.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
